Auxmos

+++ code/modules/events/spacevine.dm
@@ -326,7 +326,7 @@
 		COMSIG_ATOM_ENTERED = PROC_REF(on_entered),
 	)
 	AddElement(/datum/element/connect_loc, loc_connections)
-	AddElement(/datum/element/atmos_sensitive, mapload)
+	//AddElement(/datum/element/atmos_sensitive, mapload)
 
 /obj/structure/spacevine/examine(mob/user)
 	. = ..()
@@ -562,15 +562,13 @@
 	if(!i && prob(34 * severity))
 		qdel(src)
 
-/obj/structure/spacevine/should_atmos_process(datum/gas_mixture/air, exposed_temperature)
-	return exposed_temperature > FIRE_MINIMUM_TEMPERATURE_TO_SPREAD //if you're cold you're safe
-
-/obj/structure/spacevine/atmos_expose(datum/gas_mixture/air, exposed_temperature)
-	var/volume = air.return_volume()
-	for(var/datum/spacevine_mutation/SM in mutations)
-		if(SM.process_temperature(src, exposed_temperature, volume)) //If it's ever true we're safe
-			return
-	qdel(src)
+/obj/structure/spacevine/temperature_expose(datum/gas_mixture/air, exposed_temperature, exposed_volume)
+	if(exposed_temperature > FIRE_MINIMUM_TEMPERATURE_TO_SPREAD) //if you're cold you're safe
+		var/volume = air.return_volume()
+		for(var/datum/spacevine_mutation/SM in mutations)
+			if(SM.process_temperature(src, exposed_temperature, volume)) //If it's ever true we're safe
+				return
+		qdel(src)
 
 /obj/structure/spacevine/CanAllowThrough(atom/movable/mover, border_dir)
 	. = ..()
